+++ notification_sink/rgb_led_sink.py
@@ -11,9 +11,11 @@ from PIL import Image
 import requests
 import numpy as np
 from io import BytesIO
+from display_on_matrix.display_logic import display_task
+from display_on_matrix.text.flow import flow_text
 from helper_functions.arguments import get_arguments, get_speed, get_color
+from helper_functions.event_list import build_event_list, run_event_list
 from helper_functions.rules import get_rules
-from helper_functions.compare_time import evaluate_event_timing
 
 class RGBSink():
     def __init__(self):
@@ -47,69 +49,13 @@ class RGBSink():
             self.task = None
 
     async def build_event_list(self):
-        print("in build")
-        #severity_dict, systems_dict, argument_dict = self.get_rules(self.rules)
-        while True:
-            event_json = await self.queue.get()
-            print(f"Processing event: {event_json}")##########
-            event_args = get_arguments(self.rules, event_json["event"])
-            if event_args:
-                self.pending_events += event_args
-            print("\npending events")
-            print(self.pending_events)
-            print("\n\n")
+        await build_event_list(self.queue, self.rules, self.pending_events)
 
     async def run_event_list(self):
-        while True:
-            print(f"evaluating pending events, {len(self.pending_events)} left")
-            if self.pending_events:
-                print("Popping event")
-                event_args = self.pending_events.pop(0)
-                print(f"Popped event: {event_args}")
-                time_difference, log, event_expired = evaluate_event_timing(event_args, 120)
-                print(log)
-                print(f"The time difference is: {time_difference}")
-                if event_expired:
-                    print(f"Running event: {event_args}")
-                    await self.display_task(event_args)
-                else:
-                    print("Events expired not showing")
-
-            else:
-                print("No events found, sleeping")
-                await asyncio.sleep(1)
-                
+        await run_event_list(self.pending_events, self.display_task)
 
     async def display_task(self, event):
-        print("Running event: " + str(event))
-
-        if event["mode"].strip().lower()=="static":
-            self.animation_task = asyncio.create_task(self.static_text(event.get("args",{}),ticker=False))
-
-        elif event["mode"].strip().lower()=="text":
-            self.animation_task = asyncio.create_task(self.flow_text(event.get("args",{}),ticker=False))
-
-        elif event["mode"].strip().lower()=="ticker":
-            self.animation_task = asyncio.create_task(self.flow_text(event.get("args",{}),ticker=True))
-
-        elif event["mode"].strip().lower()=="image":
-            self.animation_task = asyncio.create_task(self.display_image(event.get("args",{})))
-
-        elif event["mode"].strip().lower()=="gif":
-            self.animation_task = asyncio.create_task(self.display_gif(event.get("args",{})))
-
-
-        if self.animation_task != None:
-            if self.flag_infinity:
-                self.animation_task.cancel()
-                self.flag_infinity = False
-            else:
-                result = await self.animation_task
-        #if selfs.animation_task != None and self.flag_infinity:
-
-        self.queue.task_done()
-            
-        self.event_args=None
+        await display_task(self, event)
 
     async def run(self):
         try:
@@ -125,9 +71,6 @@ class RGBSink():
             #await self.run_event_list()
             print("Im outside")
             return
-
-                 
-
         except asyncio.CancelledError:
             print("Cancelled Error within RGBSink")
             return
@@ -136,66 +79,28 @@ class RGBSink():
         print(f"Consuming event: {data}")#########
         await self.queue.put(data)
 
-#    def get_speed(self, speed):
-#        sleep_dict = {"5": 0.01, "4": 0.03, "3": 0.05, "2":0.1, "1":0.2}
-#        return sleep_dict[str(speed)]
-#
-#    def get_color(self, color):
-#        color = color.replace("(","").replace(")","").replace(" ","")
-#        color = color.split(",")
-#        return color 
-
-    async def flow_text(self, args, ticker = True): #, loops=5, font_sytle="./spleen-32x64.bdf"):
-        text = args.get("text")
-        font_style = args.get("font", "quattrocento48.bdf")
-        color = get_color(args.get("color","255,192,203"))
-        speed = get_speed(args.get("speed","5"))
-        loops =  int(args.get("loops", 1))
-        show_time = int(args.get("duration", 5))
-
-        print(f"Animating '{text}'\n")
-        loop = asyncio.get_running_loop()
-
-        offscreen_canvas = self.matrix.CreateFrameCanvas()
-        
-        font = graphics.Font()
-        font.LoadFont("fonts/" + font_style)
-
-        if font.baseline <= offscreen_canvas.height:
-            #Center the text vertically
-            pos_y = int(font.baseline - (offscreen_canvas.height/2 - font.baseline/2))
-        else:
-            pos_y = int(font.baseline + (offscreen_canvas.height/2 - font.baseline/2))
-
-        textcolor = graphics.Color(int(color[0]),int(color[1]),int(color[2]))
-        
-        pos_x = offscreen_canvas.width
-        
-        if ticker:
-            await self.ticker_text(offscreen_canvas, font, pos_x, pos_y, textcolor, text, loop, loops, speed)
-        else:
-            await self.timed_text(offscreen_canvas, font, pos_x, pos_y, textcolor, text, loop, show_time, speed)
-
-        return
+    # Unused/untested for now 05.12.24
+    async def flow_text(self, args, ticker=True):
+        await flow_text(self, args, ticker)
     
-    async def ticker_text(self, offscreen_canvas, font, pos_x, pos_y, textcolor, text, loop, loops, speed):
-        try:
-            while loops > 0:
-                offscreen_canvas.Clear()
-                length = graphics.DrawText(offscreen_canvas, font, pos_x, pos_y, textcolor, text)
-                pos_x -= 1
-
-                if (pos_x + length < 0):
-                    loops -= 1
-                    pos_x = offscreen_canvas.width
-
-                offscreen_canvas = await loop.run_in_executor(None, self.matrix.SwapOnVSync, offscreen_canvas)
-                # adapt to modify speed
-                await asyncio.sleep(speed)
-
-        except asyncio.CancelledError:
-            print("cancelado")
-            return "error"
+#    async def ticker_text(self, offscreen_canvas, font, pos_x, pos_y, textcolor, text, loop, loops, speed):
+#        try:
+#            while loops > 0:
+#                offscreen_canvas.Clear()
+#                length = graphics.DrawText(offscreen_canvas, font, pos_x, pos_y, textcolor, text)
+#                pos_x -= 1
+#
+#                if (pos_x + length < 0):
+#                    loops -= 1
+#                    pos_x = offscreen_canvas.width
+#
+#                offscreen_canvas = await loop.run_in_executor(None, self.matrix.SwapOnVSync, offscreen_canvas)
+#                # adapt to modify speed
+#                await asyncio.sleep(speed)
+#
+#        except asyncio.CancelledError:
+#            print("cancelado")
+#            return "error"
 
     async def static_text(self, args, ticker = True): #, loops=5, font_sytle="./spleen-32x64.bdf"):
         print("SHOWING STATIC TEXT")
